Refactor: Mejora en la organización de datos y la gestión de la puntuación más alta.
