+++ app/src/main/java/net/oscar/simondice/Datos/Datos.kt
@@ -1,9 +1,7 @@
-package net.oscar.simondice
+package net.oscar.simondice.Datos
 
 import androidx.compose.ui.graphics.Color
-import java.time.LocalTime
 
-data class PuntuacionMasAlta(val marcaTiempo: LocalTime, val record: Int)
 object Datos {
     val secuenciaAdivinar = listOf(Colores.ROJO,Colores.AZUL,Colores.AMARILLO,Colores.VERDE)
     val secuenciaAdivinando = mutableListOf<Colores>()
