+++ app/src/main/java/net/oscar/simondice/Datos/PuntuacionMasAlta.kt
@@ -1,0 +1,10 @@
+package net.oscar.simondice.Datos
+
+import android.text.format.DateUtils
+import java.time.LocalTime
+import java.util.Date
+
+object PuntuacionMasAlta {
+    val puntuacionMasAlta: Int = 0
+    val marcaTiempo: Date = Date()
+}
