+++ app/src/main/java/net/oscar/simondice/puntuacionMasAlta/PuntuacionMasAltaHandler.kt
@@ -1,0 +1,12 @@
+package net.oscar.simondice.puntuacionMasAlta
+
+import net.oscar.simondice.Datos.PuntuacionMasAlta
+
+/**
+ * Guardar los records en el juego del simon dice
+ */
+interface PuntuacionMasAltaHandler {
+    fun obtenerRecord(): PuntuacionMasAlta;
+    fun anadirRecord(puntuacionMasAlta: PuntuacionMasAlta);
+    fun eliminarRecord(puntuacionMasAlta: PuntuacionMasAlta);
+}
